+++ appl_file_std_mgr.cpp
@@ -38,9 +38,8 @@ enum appl_status
     e_status =
         appl_object::create_instance(
             p_context,
-            static_cast<unsigned long int>(
-                sizeof(
-                    class appl_file_std_mgr)),
+            sizeof(
+                class appl_file_std_mgr),
             &(
                 appl_file_std_mgr::placement_new),
             0,
